PUBDEV-2851: Add a new parameter 'max_delta_step' to GBM. Code cleanup: Modify some parameters from float to double.

+++ h2o-algos/src/main/java/hex/schemas/SharedTreeV3.java
@@ -71,16 +71,16 @@ public class SharedTreeV3<B extends SharedTree, S extends SharedTreeV3<B,S,P>, P
     public boolean build_tree_one_node;
 
     @API(help = "Row sample rate per tree (from 0.0 to 1.0)", gridable = true)
-    public float sample_rate;
+    public double sample_rate;
 
     @API(help = "Row sample rate per tree per class (from 0.0 to 1.0)", level = API.Level.expert, gridable = true)
-    public float[] sample_rate_per_class;
+    public double[] sample_rate_per_class;
 
     @API(help = "Column sample rate per tree (from 0.0 to 1.0)", level = API.Level.secondary, gridable = true)
-    public float col_sample_rate_per_tree;
+    public double col_sample_rate_per_tree;
 
     @API(help = "Relative change of the column sampling rate for every level (from 0.0 to 2.0)", level = API.Level.expert, gridable = true)
-    public float col_sample_rate_change_per_level;
+    public double col_sample_rate_change_per_level;
 
     @API(help="Score the model after every so many trees. Disabled if set to 0.", level = API.Level.secondary, gridable = false)
     public int score_tree_interval;
